remove factory file
